changed delay to be an int

+++ SimpleFractals/java/CircleLimitSet1.java
@@ -24,7 +24,7 @@ public class CircleLimitSet1 extends SimpleFractal {
 		return 36;
 	}
 
-	public long getSuggestedDelay() {
+	public int getSuggestedDelay() {
 		return 100;
 	}
